fix: ./dev.sh docker-update
We could not update npm due to the following error when running Node 12.x:

ERROR: npm v9.1.2 is known not to run on Node.js v12.22.12. You'll need to
upgrade to a newer Node.js version in order to use this version of npm. This
version of npm supports the following node versions: `^14.17.0 || ^16.13.0 ||
>=18.0.0`. You can find the latest version at https://nodejs.org/.

We could not update lerna due to the following error when running Node 12.x:

/opt/hostedtoolcache/node/12.22.12/x64/lib/node_modules/lerna/node_modules/@lerna/package-graph/index.js:79
            if (depNode?.version) {
                        ^

We could not update jsdom due to the following error when running Node 12.x:

/home/runner/work/utiljs/utiljs/node_modules/jsdom/lib/jsdom/browser/parser/html.js:170
    after._pushedOnStackOfOpenElements?.();

                                       ^